+++ hutool-cron/src/test/java/cn/hutool/cron/demo/CronTest.java
@@ -36,13 +36,14 @@ public class CronTest {
 	}
 	
 	@Test
-	@Ignore
+//	@Ignore
 	public void cronTest2() {
 		// 支持秒级别定时任务
 		CronUtil.setMatchSecond(true);
 		CronUtil.start();
-		
-		ThreadUtil.sleep(30000);
+
+		ThreadUtil.waitForDie();
+		Console.log("Exit.");
 	}
 
 	@Test